refactor: move PlaceAtDistanceListContainer to departure
