+++ app/component/route/NearbyRouteListContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Relay from 'react-relay';
-import PlaceAtDistanceListContainer from './PlaceAtDistanceListContainer';
+import PlaceAtDistanceListContainer from '../departure/PlaceAtDistanceListContainer';
 import config from '../../config';
 
 const NearbyRouteList = (props) => (
